fix bookObj type
